+++ api/department/update.php
@@ -1,0 +1,49 @@
+<?php
+    // required headers
+    header("Access-Control-Allow-Origin: *");
+    header("Content-Type: application/json; charset=UTF-8");
+    header("Access-Control-Allow-Methods: POST");
+    header("Access-Control-Max-Age: 3600");
+    header("Access-Control-Allow-Headers: Content-Type, Access-Control-Allow-Headers, Authorization, X-Requested-With");
+  
+    // include database and object files
+    include_once '../config/database.php';
+    include_once '../objects/departments.php';
+  
+    // get database connection
+    $database = new Database();
+    $db = $database->getConnection();
+  
+    // prepare product object
+    $department = new Department($db);
+  
+    // get id of product to be edited
+    $data = json_decode(file_get_contents("php://input"));
+  
+    
+    
+    // set product property values
+    $department->Name = $data->Name;
+    $department->H_UCID = $data->H_UCID;
+
+  
+    // update the product
+    if($department->update()){
+  
+        // set response code - 200 ok
+        http_response_code(200);
+  
+        // tell the user
+        echo json_encode(array("message" => "Department was updated."));
+    }
+  
+    // if unable to update the product, tell the user
+    else{
+  
+        // set response code - 503 service unavailable
+        http_response_code(503);
+  
+        // tell the user
+        echo json_encode(array("message" => "Unable to update department."));
+    }
+?>